@@ -1,8 +1,16 @@
-module.exports = {
-    Branches:                    require('./Branches.jsx'),
-    PullRequests:                require('./PullRequests.jsx'),
-    IssueLabelsTreemap:          require('./IssueLabelsTreemap.jsx'),
-    UserBadge:                   require('./UserBadge.jsx'),
-    RepositoryContributorsStats: require('./RepositoryContributorsStats.jsx'),
-    Status:                      require('./Status.jsx')
-};
+import Branches                    from './Branches.jsx';
+import PullRequests                from './PullRequests.jsx';
+import IssueLabelsTreemap          from './IssueLabelsTreemap.jsx';
+import UserBadge                   from './UserBadge.jsx';
+import RepositoryContributorsStats from './RepositoryContributorsStats.jsx';
+import Status                      from './Status.jsx';
+
+
+export default {
+    Branches:                    Branches,
+    PullRequests:                PullRequests,
+    IssueLabelsTreemap:          IssueLabelsTreemap,
+    UserBadge:                   UserBadge,
+    RepositoryContributorsStats: RepositoryContributorsStats,
+    Status:                      Status
+}
